Created generic Singleton script
This singleton script is a generic MonoBehaviour class.
It allows the user to make any class into a singleton monobehaviour by inehriting from it.

There are 2 types of singletons:
Singleton<T> : regular singleton.
SingletonPersistent<T> : same as a regular singleton, but makes use of unity specific functionality to prevent being unloaded when the scene changes.