Changed camera look and feel
Changed camera FOV from 60 to 90 to give the player a wider view of the environment. This greatly improves the experience by allowing players to see obstacles coming from far.

The camera socket has been moved upward 1 unit. This makes it so that the players will always have more space in view ahead of their ship rather than seeing more on the back of the vehicle. This increases the range of view in the direction in which the main action takes place, which also improves playeability a lot.