Createed DebugManager
The debug manager will allow setting globally whether debug "cheats" and other debug actions are enabled or not without having to rely on multiple isDebug control variables or having to constantly comment out code.

This also makes it a non-problem to forget to remove debug code from the final build since it will be disabled with a simple switch of a boolean variable. (well... kinda... it is a problem, it's just not going to break anything or lower performance since we basically are preventing any of the debug code from running during runtime in final builds by making sure that the debug flag is off. The code is still there, and a savvy user could still enable it, but let's let the players have some fun with the cheats, ok? it ain't like this is a big project or anything, so no harm done in having a literal god mode key.)