Modified player controller input checks
The player controller's inputs now make sure to check whether the player is alive or not. That way, the player can only move and shoot if they are alive.

The inputs now also check whether the player is in debug mode or not before allowing the user to execute any debug actions / cheats such as adding and removing health.