Modified Debug Manager
The debug manager allows setting the debug value from the editor on startup on the init scene. The real debug value is stored as a static variable. This means that it can be accessed without having to always check whether the DebugManager.Instance is null or not (a condition that should never happen during regular gameplay but is very common during testing).

This also means that testing is now easier, since the value is directly initialized and no exceptions can happen if we forget to check whether the instance is valid or not when starting during testing on scenes other than the init scene.