Fixed warning in init scene
The singleton persistent monobehaviour scripts attached to the persistent managers only works properly in unity if the object is a root object. These managers were attached to an empty object called "PersistentManagers" for organization purposes, but this is apparently not valid in unity.

Things still work properly, it just gives a warning, but it is better to be safe than sorry and rely only on valid and documented behaviour.