Reworked player references on HUDs
The HUD controllers no longer contain their own player controller reference. They now make use of a global reference stored within the HUDManager.Instance.

This prevents code duplication, and it makes things easier, since now we don't rely of having a set player reference function for every single HUD, reducing the amount of points in which the code could break, and also reducing the bloat and memory footprint of having tons of pointers for something that should be known at all times from the moment the scene is first loaded.