Changed obstacle movement
Obstacle movement has now been changed for a physics implementation.

Rather than changing transform values manually, the obstacles are moved using the rigid body physics. This allows for more interesting behaviours and orbits when the obstacles miss the player.

In the previous implementation, the obstacles moved faster the further they were from the player. This has been changed.

Now, the further the obstacle is located, the slower it moves toward the target (the player).

The closer the player is to the obstacle, the faster it moves.