Relocated asteroid despawn behaviour
The asteroid despawn logic is now located within a delegate that gets called onUpdate. This means that the despawn distance is no longer tied to the asteroids themselves. It is now tied to the controller class that manages the object pool that spawned the asteroid. This means that asteroids can now behave like regular objects if spawned on their own (without using a pool) without requiring any extra logic, as well as being able to control the despawn logic from within the pool object itself.