Modified obstacle scriptable object
The scriptable object for asteroid now makes use of the asteroid type enum to return data depending on what type the asteroid is.

Each asteroid / obstacle of each type has its own speed and sprites, as well as their own behaviour (asteroid type specific behaviour is to be implemented within the asteroid class itself).

The target variable has now been made private to prevent it from appearing on the editor's scriptable object menu. The target is now accessible through a getter and setter.